Politicas en vistas

+++ app/Http/Controllers/UserController.php
@@ -15,6 +15,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('haveaccess', 'user.index');
         $users = User::with('roles')->orderBy('id', 'Desc')->paginate(2);
 
         return view('user.index', compact('users'));
@@ -49,7 +50,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        $this->authorize('view', $user);
+        $this->authorize('view', [$user, ['user.show', 'userown.show']]);
         $roles = Role::orderBy('name')->get();
 
         return view('user.view', compact('roles', 'user'));
@@ -63,7 +64,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        $this->authorize('update', $user);
+        $this->authorize('update', [$user, ['user.edit', 'userown.edit']]);
         $roles = Role::orderBy('name')->get();
 
         return view('user.edit', compact('roles', 'user'));
@@ -98,6 +99,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('haveaccess', 'user.destroy');
         $user->delete();
 
         return redirect()->route('user.index')->with('status_success', 'User successfully removed');
